[TM-3107] fix minor problems for notification
