+++ .github/workflows/deploy-service.yml
@@ -85,6 +85,7 @@ jobs:
           echo "AWS_ACCESS_KEY_ID=\"${{ secrets.AWS_ACCESS_KEY_ID }}\"" >> .env
           echo "AWS_SECRET_ACCESS_KEY=\"${{ secrets.AWS_SECRET_ACCESS_KEY }}\"" >> .env
           echo "TRANSIFEX_SECRET=\"${{ secrets.TRANSIFEX_SECRET }}\"" >> .env
+          echo "GREENHOUSE_API_TOKEN=\"${{ secrets.GREENHOUSE_API_TOKEN }}\"" >> .env
           : # Don't build the base image with NODE_ENV because it'll limit the packages that are installed
           docker build -t terramatch-microservices-base:nx-base .
           SERVICE_IMAGE=$ECR_REGISTRY/$ECR_REPOSITORY:${{ env.IMAGE_TAG }}
